Updated story folder structure and changed existing experience object for Update Experience story
